get cache shouldnt call login
